Conflicts:
	lib/LuaNode/Net.lua
	src/luanode.cpp
	src/luanode_constants.cpp
	src/luanode_net.cpp
	test/simple/test-http-finish-before-resolve.lua

+++ src/luanode_crypto.h
@@ -54,7 +54,7 @@ private:
 
 private:
 	lua_State* m_L;
-	const unsigned long m_socketId;
+	const unsigned int m_socketId;
 	bool m_shutdown_pending;
 	bool m_close_pending;
 	unsigned long m_pending_writes;
