+++ src/luanode_file_linux.h
@@ -37,7 +37,7 @@ private:
 
 private:
 	lua_State* m_L;
-	const unsigned long m_fileId;
+	const unsigned int m_fileId;
 	boost::array<char, 8192> m_inputArray;	// agrandar esto y poolearlo (el test simple\test-http-upgrade-server necesita un buffer grande sino falla)
 };
